chore: fix #1474
